Update self.__session to Session

+++ models/engine/db_storage.py
@@ -2,9 +2,6 @@
 '''
     describe DBstorage
 '''
-from sqlalchemy import create_engine
-from sqlalchemy.orm import sessionmaker, scoped_session
-from sqlalchemy.ext.declarative import declarative_base
 from models.base_model import Base
 from models.state import State
 from models.city import City
@@ -12,6 +9,8 @@ from models.amenity import Amenity
 from models.place import Place
 from models.review import Review
 from models.user import User
+from sqlalchemy import create_engine
+from sqlalchemy.orm import sessionmaker, scoped_session
 import os
 
 
@@ -32,13 +31,7 @@ class DBStorage:
                                           os.environ['HBNB_MYSQL_HOST'],
                                           os.environ['HBNB_MYSQL_DB']),
                                       pool_pre_ping=True)
-        Base.metadata.create_all(self.__engine)
-
-        Session = sessionmaker(bind=self.__engine)
-        session = Session()
-
         if os.environ['HBNB_ENV'] == 'test':
-            # DROP ALL tables
             Base.metadata.drop_all(self.__engine)
 
     def all(self, cls=None):
@@ -72,39 +65,25 @@ class DBStorage:
         '''
             add object to current session.
         '''
-        Session = sessionmaker(bind=self.__engine)
-        session = Session()
         session.add(obj)
-        session.commit()
 
     def save(self):
         '''
             commit all changes to current session.
         '''
-        Session = sessionmaker(bind=self.__engine)
-        session = Session()
         session.commit()
 
     def delete(self, obj=None):
         '''
             delete from current sesssion
         '''
-        Session = sessionmaker(bind=self.__engine)
-        session = Session()
-        session.delete(obj)
-        session.commit()
+        if obj:
+            session.delete(obj)
 
     def reload(self):
         '''
           reload stuff
         '''
-        self.__engine = create_engine("mysql+mysqldb://{}:{}@{}/{}".format
-                                      (os.environ['HBNB_MYSQL_USER'],
-                                       os.environ['HBNB_MYSQL_PWD'],
-                                       os.environ['HBNB_MYSQL_HOST'],
-                                       os.environ['HBNB_MYSQL_DB']),
-                                      pool_pre_ping=True)
         Base.metadata.create_all(self.__engine)
-
         Session = sessionmaker(bind=self.__engine, expire_on_commit=False)
-        session = scoped_session(Session())
+        self.__session = scoped_session(Session)
